# Optimize conversion of KeyGesture from/to string, reduce allocations