--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Command/KeyGestureConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Command/KeyGestureConverter.cs
@@ -138,11 +138,10 @@ namespace System.Windows.Input
             // You will only get string.Empty from KeyConverter for Key.None and we've checked that above
             string strKey = (string)s_keyConverter.ConvertTo(context, culture, keyGesture.Key, destinationType);
 
-            // No modifiers, just binding (possibly with with display string)
+            // No modifiers, just binding (possibly with with display string) -> "F5,Refresh"
             if (keyGesture.Modifiers is ModifierKeys.None)
                 return string.IsNullOrEmpty(keyGesture.DisplayString) ? strKey : $"{strKey},{keyGesture.DisplayString}";
 
-            // Prepend modifiers
             ReadOnlySpan<char> modifierSpan = ModifierKeysConverter.ConvertMultipleModifiers(keyGesture.Modifiers, stackalloc char[22]);
 
             // Append display string if there's any, like "Ctrl+A,Description"